+++ optuna/storages/base.py
@@ -1,8 +1,9 @@
 import abc
 import copy
 
-from optuna import structs
 from optuna import type_checking
+from optuna.study import StudyDirection
+from optuna.trial import TrialState
 
 if type_checking.TYPE_CHECKING:
     from typing import Any  # NOQA
@@ -11,6 +12,8 @@ if type_checking.TYPE_CHECKING:
     from typing import Optional  # NOQA
 
     from optuna import distributions  # NOQA
+    from optuna.study import StudySummary  # NOQA
+    from optuna.trial import FrozenTrial  # NOQA
 
 DEFAULT_STUDY_NAME_PREFIX = "no-name-"
 
@@ -46,7 +49,7 @@ class BaseStorage(object, metaclass=abc.ABCMeta):
 
     @abc.abstractmethod
     def set_study_direction(self, study_id, direction):
-        # type: (int, study.StudyDirection) -> None
+        # type: (int, StudyDirection) -> None
 
         raise NotImplementedError
 
@@ -78,7 +81,7 @@ class BaseStorage(object, metaclass=abc.ABCMeta):
 
     @abc.abstractmethod
     def get_study_direction(self, study_id):
-        # type: (int) -> study.StudyDirection
+        # type: (int) -> StudyDirection
 
         raise NotImplementedError
 
@@ -96,7 +99,7 @@ class BaseStorage(object, metaclass=abc.ABCMeta):
 
     @abc.abstractmethod
     def get_all_study_summaries(self):
-        # type: () -> List[study.StudySummary]
+        # type: () -> List[StudySummary]
 
         raise NotImplementedError
 
@@ -104,13 +107,13 @@ class BaseStorage(object, metaclass=abc.ABCMeta):
 
     @abc.abstractmethod
     def create_new_trial(self, study_id, template_trial=None):
-        # type: (int, Optional[trial.FrozenTrial]) -> int
+        # type: (int, Optional[FrozenTrial]) -> int
 
         raise NotImplementedError
 
     @abc.abstractmethod
     def set_trial_state(self, trial_id, state):
-        # type: (int, trial.TrialState) -> bool
+        # type: (int, TrialState) -> bool
 
         raise NotImplementedError
 
@@ -160,32 +163,32 @@ class BaseStorage(object, metaclass=abc.ABCMeta):
 
     @abc.abstractmethod
     def get_trial(self, trial_id):
-        # type: (int) -> trial.FrozenTrial
+        # type: (int) -> FrozenTrial
 
         raise NotImplementedError
 
     @abc.abstractmethod
     def get_all_trials(self, study_id, deepcopy=True):
-        # type: (int, bool) -> List[trial.FrozenTrial]
+        # type: (int, bool) -> List[FrozenTrial]
 
         raise NotImplementedError
 
     @abc.abstractmethod
     def get_n_trials(self, study_id, state=None):
-        # type: (int, Optional[trial.TrialState]) -> int
+        # type: (int, Optional[TrialState]) -> int
 
         raise NotImplementedError
 
     def get_best_trial(self, study_id):
-        # type: (int) -> trial.FrozenTrial
+        # type: (int) -> FrozenTrial
 
         all_trials = self.get_all_trials(study_id, deepcopy=False)
-        all_trials = [t for t in all_trials if t.state is trial.TrialState.COMPLETE]
+        all_trials = [t for t in all_trials if t.state is TrialState.COMPLETE]
 
         if len(all_trials) == 0:
             raise ValueError("No trials are completed yet.")
 
-        if self.get_study_direction(study_id) == study.StudyDirection.MAXIMIZE:
+        if self.get_study_direction(study_id) == StudyDirection.MAXIMIZE:
             best_trial = max(all_trials, key=lambda t: t.value)
         else:
             best_trial = min(all_trials, key=lambda t: t.value)
@@ -213,7 +216,7 @@ class BaseStorage(object, metaclass=abc.ABCMeta):
         pass
 
     def check_trial_is_updatable(self, trial_id, trial_state):
-        # type: (int, trial.TrialState) -> None
+        # type: (int, TrialState) -> None
 
         if trial_state.is_finished():
             trial = self.get_trial(trial_id)
