+++ optuna/storages/in_memory.py
@@ -5,7 +5,8 @@ import threading
 from optuna import distributions  # NOQA
 from optuna.storages import base
 from optuna.storages.base import DEFAULT_STUDY_NAME_PREFIX
-from optuna import structs
+from optuna.study import StudyDirection, StudySummary
+from optuna.trial import FrozenTrial, TrialState
 from optuna import type_checking
 
 if type_checking.TYPE_CHECKING:
@@ -26,9 +27,9 @@ class InMemoryStorage(base.BaseStorage):
 
     def __init__(self):
         # type: () -> None
-        self.trials = []  # type: List[trial.FrozenTrial]
+        self.trials = []  # type: List[FrozenTrial]
         self.param_distribution = {}  # type: Dict[str, distributions.BaseDistribution]
-        self.direction = study.StudyDirection.NOT_SET
+        self.direction = StudyDirection.NOT_SET
         self.study_user_attrs = {}  # type: Dict[str, Any]
         self.study_system_attrs = {}  # type: Dict[str, Any]
         self.study_name = DEFAULT_STUDY_NAME_PREFIX + IN_MEMORY_STORAGE_STUDY_UUID  # type: str
@@ -64,16 +65,16 @@ class InMemoryStorage(base.BaseStorage):
             self.trials = []
             self.best_trial_id = None
             self.param_distribution = {}
-            self.direction = study.StudyDirection.NOT_SET
+            self.direction = StudyDirection.NOT_SET
             self.study_user_attrs = {}
             self.study_system_attrs = {}
             self.study_name = DEFAULT_STUDY_NAME_PREFIX + IN_MEMORY_STORAGE_STUDY_UUID
 
     def set_study_direction(self, study_id, direction):
-        # type: (int, study.StudyDirection) -> None
+        # type: (int, StudyDirection) -> None
 
         with self._lock:
-            if self.direction != study.StudyDirection.NOT_SET and self.direction != direction:
+            if self.direction != StudyDirection.NOT_SET and self.direction != direction:
                 raise ValueError(
                     "Cannot overwrite study direction from {} to {}.".format(
                         self.direction, direction
@@ -113,7 +114,7 @@ class InMemoryStorage(base.BaseStorage):
         return self.study_name
 
     def get_study_direction(self, study_id):
-        # type: (int) -> study.StudyDirection
+        # type: (int) -> StudyDirection
 
         return self.direction
 
@@ -130,10 +131,10 @@ class InMemoryStorage(base.BaseStorage):
             return copy.deepcopy(self.study_system_attrs)
 
     def get_all_study_summaries(self):
-        # type: () -> List[study.StudySummary]
+        # type: () -> List[StudySummary]
 
         best_trial = None
-        if any(t for t in self.trials if t.state == trial.TrialState.COMPLETE):
+        if any(t for t in self.trials if t.state == TrialState.COMPLETE):
             best_trial = self.get_best_trial(IN_MEMORY_STORAGE_STUDY_ID)
 
         datetime_start = None
@@ -141,7 +142,7 @@ class InMemoryStorage(base.BaseStorage):
             datetime_start = min([t.datetime_start for t in self.trials])
 
         return [
-            study.StudySummary(
+            StudySummary(
                 study_name=self.study_name,
                 direction=self.direction,
                 best_trial=best_trial,
@@ -154,7 +155,7 @@ class InMemoryStorage(base.BaseStorage):
         ]
 
     def create_new_trial(self, study_id, template_trial=None):
-        # type: (int, Optional[trial.FrozenTrial]) -> int
+        # type: (int, Optional[FrozenTrial]) -> int
 
         self._check_study_id(study_id)
 
@@ -173,12 +174,12 @@ class InMemoryStorage(base.BaseStorage):
 
     @staticmethod
     def _create_running_trial():
-        # type: () -> trial.FrozenTrial
+        # type: () -> FrozenTrial
 
-        return trial.FrozenTrial(
+        return FrozenTrial(
             trial_id=-1,  # dummy value.
             number=-1,  # dummy value.
-            state=trial.TrialState.RUNNING,
+            state=TrialState.RUNNING,
             params={},
             distributions={},
             user_attrs={},
@@ -190,13 +191,13 @@ class InMemoryStorage(base.BaseStorage):
         )
 
     def set_trial_state(self, trial_id, state):
-        # type: (int, trial.TrialState) -> bool
+        # type: (int, TrialState) -> bool
 
         with self._lock:
             trial = self.trials[trial_id]
             self.check_trial_is_updatable(trial_id, trial.state)
 
-            if state == trial.TrialState.RUNNING and trial.state != trial.TrialState.WAITING:
+            if state == TrialState.RUNNING and trial.state != TrialState.WAITING:
                 return False
 
             trial.state = state
@@ -239,7 +240,7 @@ class InMemoryStorage(base.BaseStorage):
         return trial_id
 
     def get_best_trial(self, study_id):
-        # type: (int) -> trial.FrozenTrial
+        # type: (int) -> FrozenTrial
 
         if self.best_trial_id is None:
             raise ValueError("No trials are completed yet.")
@@ -263,7 +264,7 @@ class InMemoryStorage(base.BaseStorage):
     def _update_cache(self, trial_id):
         # type: (int) -> None
 
-        if self.trials[trial_id].state != trial.TrialState.COMPLETE:
+        if self.trials[trial_id].state != TrialState.COMPLETE:
             return
 
         if self.best_trial_id is None:
@@ -277,7 +278,7 @@ class InMemoryStorage(base.BaseStorage):
         # Complete trials do not have `None` values.
         assert new_value is not None
 
-        if self.get_study_direction(IN_MEMORY_STORAGE_STUDY_ID) == study.StudyDirection.MAXIMIZE:
+        if self.get_study_direction(IN_MEMORY_STORAGE_STUDY_ID) == StudyDirection.MAXIMIZE:
             if best_value < new_value:
                 self.best_trial_id = trial_id
             return
@@ -316,13 +317,13 @@ class InMemoryStorage(base.BaseStorage):
             self.trials[trial_id].system_attrs[key] = value
 
     def get_trial(self, trial_id):
-        # type: (int) -> trial.FrozenTrial
+        # type: (int) -> FrozenTrial
 
         with self._lock:
             return copy.deepcopy(self.trials[trial_id])
 
     def get_all_trials(self, study_id, deepcopy=True):
-        # type: (int, bool) -> List[trial.FrozenTrial]
+        # type: (int, bool) -> List[FrozenTrial]
 
         self._check_study_id(study_id)
         with self._lock:
@@ -332,7 +333,7 @@ class InMemoryStorage(base.BaseStorage):
                 return self.trials
 
     def get_n_trials(self, study_id, state=None):
-        # type: (int, Optional[trial.TrialState]) -> int
+        # type: (int, Optional[TrialState]) -> int
 
         self._check_study_id(study_id)
         if state is None:
