+++ optuna/study/__init__.py
@@ -1,0 +1,3 @@
+from optuna.study import study  # NOQA
+from optuna.study.study_direction import StudyDirection  # NOQA
+from optuna.study.study_summary import StudySummary  # NOQA
